Fixed all the required code smells from the implementation and design set
